third_party/WebKit/LayoutTests/external/wpt/async_cookies/observation_static.tentative-expected.txt

This is a testharness.js-based test.
FAIL Cookie Store Tests (testObservation) assert_equals: testDeleteCookies failures: TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function,TypeError: cookieStore.delete is not a function expected 0 but got 12
Harness: the test ran to completion.
